requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-allauth==65.4.1
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.1
httplib2==0.22.0
idna==3.10
multidict==6.1.0
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
twilio==9.4.6
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
